+++ src/types.ts
@@ -1,0 +1,13 @@
+import { keywords } from "./constants";
+
+export type Keyword = (typeof keywords)[number];
+
+export type Token =
+  | {
+      type: "keyword";
+      value: Keyword;
+    }
+  | {
+      type: "string";
+      value: string;
+    };
